Pull out status_t to the top level to allow using those enum values from clients.
